+++ changelog.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+### v0.3 (9/October/2020)
+
+- Renamed `lib` table to `package`, But you can still use `lib` as cherry is backward-compatible!
+
+> NOTE 1: Compatibility was made from this version so it can run packages that were created with old versions of cherry.
+
 ### v0.2 (8/October/2020)
 
 - Add possibility of folders and branching!